Move testing extension into features folder
