+++ behat.yml
@@ -13,4 +13,4 @@ default:
           selenium2:
             browser: chrome
             wd_host: http://dockermachine.local:4444/wd/hub
-    Behat\FlexibleMink\Extension\Assertion\AssertionExtension: ~
+    features\Extensions\Assertion\AssertionExtension: ~
