+++ composer.json
@@ -14,7 +14,8 @@
   },
   "autoload": {
     "psr-0": {
-      "Behat\\FlexibleMink\\": "src/"
+      "Behat\\FlexibleMink\\": "src/",
+      "features\\": ""
     }
   }
 }
